+++ ml_dtypes/tests/float8_test.cc
@@ -40,6 +40,8 @@ struct Float8TestParamNames {
       return "float8_e4m3fn";
     } else if constexpr (std::is_same_v<TypeParam, float8_e4m3b11fnuz>) {
       return "float8_e4m3b11fnuz";
+    } else if constexpr (std::is_same_v<TypeParam, float8_e3m4>) {
+      return "float8_e3m4";
     } else if constexpr (std::is_same_v<TypeParam, float8_e4m3>) {
       return "float8_e4m3";
     } else if constexpr (std::is_same_v<TypeParam, float8_e5m2>) {
@@ -54,9 +56,40 @@ struct Float8TestParamNames {
 };
 
 using Float8Types =
-    ::testing::Types<float8_e4m3, float8_e4m3fn, float8_e5m2,
+    ::testing::Types<float8_e3m4, float8_e4m3, float8_e4m3fn, float8_e5m2,
                      float8_e4m3b11fnuz, float8_e4m3fnuz, float8_e5m2fnuz>;
 TYPED_TEST_SUITE(Float8Test, Float8Types, Float8TestParamNames);
+
+TEST(Float8E3m4Test, NumericLimits) {
+  EXPECT_TRUE(
+      Eigen::numext::isnan(std::numeric_limits<float8_e3m4>::quiet_NaN()));
+  EXPECT_TRUE(
+      Eigen::numext::isnan(std::numeric_limits<float8_e3m4>::signaling_NaN()));
+  EXPECT_EQ(static_cast<float>(std::numeric_limits<float8_e3m4>::min()),
+            0.25);
+  EXPECT_EQ(static_cast<float>(std::numeric_limits<float8_e3m4>::max()), 15.5);
+  EXPECT_EQ(static_cast<float>(std::numeric_limits<float8_e3m4>::lowest()),
+            -15.5);
+  EXPECT_EQ(static_cast<float>(std::numeric_limits<float8_e3m4>::epsilon()),
+            0.0625);
+  EXPECT_EQ(static_cast<float>(std::numeric_limits<float8_e3m4>::round_error()),
+            0.5);
+  EXPECT_TRUE(
+      Eigen::numext::isinf(std::numeric_limits<float8_e3m4>::infinity()));
+  EXPECT_EQ(static_cast<float>(std::numeric_limits<float8_e3m4>::denorm_min()),
+            std::exp2(-6));
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::digits, 5);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::digits10, 1);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::max_digits10, 3);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::min_exponent, -1);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::min_exponent10, 0);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::max_exponent, 4);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::max_exponent10, 1);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::is_iec559, true);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::has_infinity, true);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::has_quiet_NaN, true);
+  EXPECT_EQ(std::numeric_limits<float8_e3m4>::has_signaling_NaN, true);
+}
 
 TEST(Float8E4m3Test, NumericLimits) {
   EXPECT_TRUE(
@@ -610,6 +643,20 @@ TEST(Float8Test, Float8E4m3b11fnuz_To_Float8E4m3fn) {
   }
 }
 
+TEST(Float8Test, Float8E3m4_To_Float8E5m2) {
+  // Truncation and rounding of a number ever-so-slightly less than 2.
+  float8_e3m4 less_than_two = float8_e3m4::FromRep(0x3F);
+  float8_e5m2 truncated =
+      float8_e5m2::template ConvertFrom</*kSaturate=*/false,
+                                        /*kTruncate=*/true>(less_than_two);
+  EXPECT_LT(static_cast<float>(truncated), 2);
+
+  float8_e5m2 rounded =
+      float8_e5m2::template ConvertFrom</*kSaturate=*/false,
+                                        /*kTruncate=*/false>(less_than_two);
+  EXPECT_EQ(static_cast<float>(rounded), 2);
+}
+
 TEST(Float8Test, Float8E4m3_To_Float8E5m2) {
   // Truncation and rounding of a number ever-so-slightly less than 2.
   float8_e4m3 less_than_two = float8_e4m3::FromRep(0x3F);
@@ -636,6 +683,67 @@ TEST(Float8Test, Float8E4m3fn_To_Float8E5m2) {
       float8_e5m2::template ConvertFrom</*kSaturate=*/false,
                                         /*kTruncate=*/false>(less_than_two);
   EXPECT_EQ(static_cast<float>(rounded), 2);
+}
+
+TEST(Float8Test, Half_To_Float8E3m4) {
+  // Special values, NaN.
+  Eigen::half inf =
+      Eigen::numext::bit_cast<Eigen::half>(static_cast<uint16_t>(0x7C00));
+  EXPECT_EQ(static_cast<float8_e3m4>(inf).rep(), 0x70);
+  Eigen::half ninf =
+      Eigen::numext::bit_cast<Eigen::half>(static_cast<uint16_t>(0xFC00));
+  EXPECT_EQ(static_cast<float8_e3m4>(ninf).rep(), 0xF0);
+
+  Eigen::half nan =
+      Eigen::numext::bit_cast<Eigen::half>(static_cast<uint16_t>(0x7C01));
+  EXPECT_EQ(static_cast<float8_e3m4>(nan).rep(), 0x78);
+  Eigen::half nnan =
+      Eigen::numext::bit_cast<Eigen::half>(static_cast<uint16_t>(0xFC01));
+  EXPECT_EQ(static_cast<float8_e3m4>(nnan).rep(), 0xF8);
+
+  // Rounding vs truncation.
+  Eigen::half less_than_two =
+      Eigen::numext::bit_cast<Eigen::half>(static_cast<uint16_t>(0x3FFF));
+  EXPECT_EQ((float8_e3m4::ConvertFrom</*kSaturate=*/false,
+                                      /*kTruncate=*/false>(less_than_two)
+                 .rep()),
+            0x40);
+  EXPECT_EQ((float8_e3m4::ConvertFrom</*kSaturate=*/false,
+                                      /*kTruncate=*/true>(less_than_two)
+                 .rep()),
+            0x3F);
+  EXPECT_EQ((float8_e3m4::ConvertFrom</*kSaturate=*/false,
+                                      /*kTruncate=*/false>(-less_than_two)
+                 .rep()),
+            0xC0);
+  EXPECT_EQ((float8_e3m4::ConvertFrom</*kSaturate=*/false,
+                                      /*kTruncate=*/true>(-less_than_two)
+                 .rep()),
+            0xBF);
+
+  // Saturation.
+  // f8e3m4<max>=0.110.1111 0x1.Fp+3 f16=0.10010.1111000000 uint16=0x4BC0
+  // f8e3m4<inf>=0.111.0000 0x1.0p+4 f16=0.10011.0000000000 uint16=0x4C00
+  for (uint16_t i = 0x4BC0; i < 0x4C00; ++i) {
+    Eigen::half big_half = Eigen::numext::bit_cast<Eigen::half>(i);
+    float big_float = static_cast<float>(big_half);
+    EXPECT_EQ(
+        (float8_e3m4::ConvertFrom</*kSaturate=*/true, /*kTruncate=*/false>(
+             big_half)
+             .rep()),
+        (float8_e3m4::ConvertFrom</*kSaturate=*/true, /*kTruncate=*/false>(
+             big_float)
+             .rep()))
+        << i;
+    EXPECT_EQ(
+        (float8_e3m4::ConvertFrom</*kSaturate=*/true, /*kTruncate=*/false>(
+             -big_half)
+             .rep()),
+        (float8_e3m4::ConvertFrom</*kSaturate=*/true, /*kTruncate=*/false>(
+             -big_float)
+             .rep()))
+        << i;
+  }
 }
 
 TEST(Float8Test, Half_To_Float8E4m3) {
@@ -831,22 +939,31 @@ TYPED_TEST(Float8Test, CallTheConstOperator) {
   }
 }
 
+TEST(Float8E3m4Test, SmallCastToDenormal) {
+  // Special edge-case where rounding to a normalized value would
+  // normally round down, but rounding to a subnormal rounds up.
+  float x = 0x0.8Ap-2;  // btw denormals
+  float8_e3m4 y = static_cast<float8_e3m4>(x);
+  float z = static_cast<float>(y);
+  EXPECT_EQ(z, 0x0.9p-2);  // rounded up to the next denormal
+}
+
 TEST(Float8E4m3Test, SmallCastToDenormal) {
   // Special edge-case where rounding to a normalized value would
   // normally round down, but rounding to a subnormal rounds up.
-  float x = std::ldexp(1.3125, -8);
+  float x = 0x0.94p-6;  // btw denormals
   float8_e4m3 y = static_cast<float8_e4m3>(x);
   float z = static_cast<float>(y);
-  EXPECT_EQ(z, std::ldexp(1.5, -8));
+  EXPECT_EQ(z, 0x0.Ap-6);  // rounded up to the next denormal
 }
 
 TEST(Float8E5m2Test, SmallCastToDenormal) {
   // Special edge-case where rounding to a normalized value would
   // normally round down, but rounding to a subnormal rounds up.
-  float x = std::ldexp(1.3125, -15);
+  float x = 0x0.A8p-14;  // btw denormals
   float8_e5m2 y = static_cast<float8_e5m2>(x);
   float z = static_cast<float>(y);
-  EXPECT_EQ(z, std::ldexp(1.5, -15));
+  EXPECT_EQ(z, 0x0.Cp-14);  // rounded up to the next denormal
 }
 
 // Helper utility for prettier test names.
@@ -872,16 +989,17 @@ struct Float8CastTestParamNames {
   GEN_LONG_DOUBLE_PAIR(Type)                                               \
   std::pair<Type, double>, std::pair<Type, float>,                         \
       std::pair<Type, Eigen::bfloat16>, std::pair<Type, Eigen::half>,      \
-      std::pair<Type, float8_e4m3>,                                        \
+      std::pair<Type, float8_e3m4>, std::pair<Type, float8_e4m3>,          \
       std::pair<Type, float8_e4m3fn>, std::pair<Type, float8_e4m3b11fnuz>, \
       std::pair<Type, float8_e4m3fnuz>, std::pair<Type, float8_e5m2fnuz>,  \
       std::pair<Type, float8_e5m2>, std::pair<Type, bool>,                 \
       std::pair<Type, int32_t>, std::pair<Type, int64_t>
 
-#define GEN_TYPE_PAIRS()                                             \
-  GEN_DEST_TYPES(float8_e4m3fn), GEN_DEST_TYPES(float8_e4m3b11fnuz), \
-      GEN_DEST_TYPES(float8_e5m2), GEN_DEST_TYPES(float8_e4m3fnuz),  \
-      GEN_DEST_TYPES(float8_e5m2fnuz), GEN_DEST_TYPES(float8_e4m3)
+#define GEN_TYPE_PAIRS()                                                       \
+  GEN_DEST_TYPES(float8_e3m4), GEN_DEST_TYPES(float8_e4m3),                    \
+      GEN_DEST_TYPES(float8_e4m3fn), GEN_DEST_TYPES(float8_e4m3b11fnuz),       \
+      GEN_DEST_TYPES(float8_e5m2), GEN_DEST_TYPES(float8_e4m3fnuz),            \
+      GEN_DEST_TYPES(float8_e5m2fnuz)
 
 using Float8CastTypePairs = ::testing::Types<GEN_TYPE_PAIRS()>;