+++ CHANGELOG.md
@@ -23,8 +23,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
-* Added new 8-bit float type following IEEE 754 convention:
-  `ml_dtypes.float8_e4m3`.
+* Added new 8-bit float types following IEEE 754 convention:
+  `ml_dtypes.float8_e4m3` and `ml_dtypes.float8_e3m4`.
 * Fix outputs of float `divmod` and `floor_divide` when denominator is zero.
 
 ## [0.4.0] - 2024-04-1